// front/src/types/Producto.ts
export interface Producto {
  _id: number;
  name: string;
  image: string;
  description: string;
  brand: string;
  category: string;
  price: number;
  countInStock: number;
  rating: number;
  numReviews: number;
}
export interface Producto {
  _id: number;
  name: string;
  brand: string;
  category: string;
  countInStock: number;
  description: string;
  image: string;
  numReviews: number;
  price: number;
  rating: number;
}